Rewrite FFI
Much has changed since 2016!

In particular, you can't use slice_from_parts -> vec to drop data that
originated in Rust and was leaked across the FFi. You must construct a
Vec directly (or a boxed slice, in our case), otherwise you're making
a copy of _data you already allocated_, which is a leak.

In order to make this distinction, we now have InternalArray, for data
that originates in Rust, and ExternalArray, for data that originates
outside Rust. Constructing Rust-native primitives works differently
for both, giving us a better chance of not accidentally writing an
unsound API / leaking memory / introducing UB